chore: split playwright.fixtures into files (3) (#3984)
