+++ package.json
@@ -50,7 +50,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^0.9.6",
-    "@playwright/test-runner": "^0.9.14",
+    "@playwright/test-runner": "^0.9.15",
     "@types/debug": "^4.1.5",
     "@types/extract-zip": "^1.6.2",
     "@types/mime": "^2.0.3",
