+++ test/keyboard.spec.ts
@@ -15,7 +15,8 @@
  * limitations under the License.
  */
 
-import { it, expect, options, attachFrame } from './playwright.fixtures';
+import { it, expect, options } from './playwright.fixtures';
+import { attachFrame } from './utils';
 
 it('should type into a textarea', async ({page}) => {
   await page.evaluate(() => {
