+++ test/network-request.spec.ts
@@ -15,7 +15,8 @@
  * limitations under the License.
  */
 
-import { it, expect, options, attachFrame } from './playwright.fixtures';
+import { it, expect, options } from './playwright.fixtures';
+import { attachFrame } from './utils';
 
 it('should work for main frame navigation request', async ({page, server}) => {
   const requests = [];
