+++ test/page-screenshot.spec.ts
@@ -15,7 +15,8 @@
  * limitations under the License.
  */
 
-import { it, expect, describe, options, verifyViewport } from './playwright.fixtures';
+import { it, expect, describe, options } from './playwright.fixtures';
+import { verifyViewport } from './utils';
 import path from 'path';
 import fs from 'fs';
