+++ test/page-wait-for-navigation.spec.ts
@@ -17,6 +17,7 @@
 
 import { it, expect } from './playwright.fixtures';
 import type { Frame } from '../index';
+import { expectedSSLError } from './utils';
 
 it('should work', async ({page, server}) => {
   await page.goto(server.EMPTY_PAGE);
@@ -70,14 +71,14 @@ it('should work with clicking on anchor links', async ({page, server}) => {
   expect(page.url()).toBe(server.EMPTY_PAGE + '#foobar');
 });
 
-it('should work with clicking on links which do not commit navigation', async ({page, server, httpsServer, expectedSSLError}) => {
+it('should work with clicking on links which do not commit navigation', async ({page, server, httpsServer, browserName}) => {
   await page.goto(server.EMPTY_PAGE);
   await page.setContent(`<a href='${httpsServer.EMPTY_PAGE}'>foobar</a>`);
   const [error] = await Promise.all([
     page.waitForNavigation().catch(e => e),
     page.click('a'),
   ]);
-  expect(error.message).toContain(expectedSSLError);
+  expect(error.message).toContain(expectedSSLError(browserName));
 });
 
 it('should work with history.pushState()', async ({page, server}) => {
