+++ test/slowmo.spec.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import { it, expect, describe, options, attachFrame } from './playwright.fixtures';
+import { it, expect, describe, options } from './playwright.fixtures';
+import { attachFrame } from './utils';
 
 async function checkSlowMo(toImpl, page, task) {
   let didSlowMo = false;
